src/main/java/com/leetcode/algorithm/medium/lowestcommonancestor/Solution.java
package com.leetcode.algorithm.medium.lowestcommonancestor;

import com.leetcode.algorithm.common.TreeNode;

class Solution {
  public TreeNode lowestCommonAncestor(TreeNode root, TreeNode p, TreeNode q) {
    if (root == null || root == p || root == q) {
      return root;
    }
    TreeNode left = lowestCommonAncestor(root.left, p, q);
    TreeNode right = lowestCommonAncestor(root.right, p, q);
    if (left != null && right != null) {
      return root;
    }
    if (left == null && right == null) {
      return null;
    }
    return left == null ? right : left;
  }
}
